separated the hi function into it's own file for readability purposes
